reconstruct slightly
